miglioramento metodo deleteProdotto

+++ src/main/java/unicam/progettofiliera/service/VenditoreHandler.java
@@ -1,11 +1,13 @@
 package unicam.progettofiliera.service;
 
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import unicam.progettofiliera.infrastructure.ApprovazioneProdottiRepository;
 import unicam.progettofiliera.infrastructure.ProdottoRepository;
 import unicam.progettofiliera.infrastructure.VenditoreRepository;
 import unicam.progettofiliera.models.prodotti.Prodotto;
+import unicam.progettofiliera.models.prodotti.ProdottoProduttore;
 import unicam.progettofiliera.models.venditori.Distributore;
 import unicam.progettofiliera.models.venditori.Produttore;
 import unicam.progettofiliera.models.venditori.Trasformatore;
@@ -20,13 +22,18 @@ public class VenditoreHandler {
     ApprovazioneProdottiRepository approvazioneProdottiRepository;
 
     public VenditoreHandler() {}
+
     @Autowired
-    public VenditoreHandler(ProdottoRepository prodottoRepository, VenditoreRepository venditoreRepository, ApprovazioneProdottiRepository approvazioneProdottiRepository) {
+    public VenditoreHandler(ProdottoRepository prodottoRepository,
+                            VenditoreRepository venditoreRepository,
+                            ApprovazioneProdottiRepository approvazioneProdottiRepository) {
         this.prodottoRepository = prodottoRepository;
         this.venditoreRepository = venditoreRepository;
+        this.approvazioneProdottiRepository = approvazioneProdottiRepository; //TODO
     }
 
-    public void caricaProdottoProduttore(String nome, Long idVenditore, String descrizione, double prezzo, String processiProduttivi){
+    public void caricaProdottoProduttore(Long idVenditore, String nome, String descrizione,
+                                         double prezzo, String processiProduttivi) {
         Venditore venditore = venditoreRepository.findById(idVenditore).
                 orElseThrow(() -> new RuntimeException("Venditore non trovato"));
         if (venditore instanceof Produttore produttore){
@@ -35,7 +42,8 @@ public class VenditoreHandler {
         } else throw new IllegalArgumentException("il venditore non è un Produttore");
     }
 
-    public void caricaProdottoTrasformatore(String nome, Long idVenditore, String descrizione, double prezzo, String processiTraformativi, List<Produttore> collaboratori){
+    public void caricaProdottoTrasformatore(Long idVenditore, String nome, String descrizione, double prezzo,
+                                            String processiTraformativi, List<Produttore> collaboratori) {
         Venditore venditore = venditoreRepository.findById(idVenditore).
                 orElseThrow(() -> new RuntimeException("Venditore non trovato"));
         if (venditore instanceof Trasformatore trasformatore){
@@ -53,10 +61,16 @@ public class VenditoreHandler {
         } else throw new IllegalArgumentException("il venditore non è un Trasformatore");
     }
 
-    public void deleteProdotto(Long idProdotto){
-        Prodotto prodotto = prodottoRepository.findById(idProdotto).orElseThrow(() -> new RuntimeException("Prodotto non trovato"));
-        prodottoRepository.delete(prodotto);
-
+    public void deleteProdotto(Long idVenditore, Long idProdotto) {
+        Venditore venditore = venditoreRepository.findById(idVenditore).
+                orElseThrow(() -> new RuntimeException("Venditore non trovato"));
+        Prodotto prodotto = prodottoRepository.findById(idProdotto).
+                orElseThrow(() -> new RuntimeException("Prodotto non trovato"));
+        if(venditore.getProdottiPubblicati().contains(prodotto)){
+            venditore.getProdottiPubblicati().remove(prodotto);
+            prodottoRepository.delete(prodotto);
+        } else
+            throw new RuntimeException("Il prodotto non appartiene al venditore selezionato");
     }
 
 }
